+++ Scenes/World/stats_panel.gd
@@ -5,6 +5,6 @@ extends PanelContainer
 @onready var z = $StatsContainer/Distance
 
 func run():
-	x.Display = RunHandler.Mats
+	x.Display = RunHandler.GetTotalResources()
 	y.text = "Time Spent: " + String.num(RunHandler.TimeSpent, 1) + " seconds"
 	z.text = "Distance Covered: " + String.num(RunHandler.DistanceTravelled / 16.0, 1) + "u"