+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added Bigger Bay upgrade to Inventory Monitor
 - Added BERTHA upgrade to Bigger Bay
 - Added Compartmentalization upgrade to Bigger Bay
+- Added Mass Driver upgrade to Compartmentalization
+- Added Borealis Station upgrade to Core Assembler
 - Added flare shader to Apollo (replaces inner ring)
 - Reduced rotation speed of Apollo, and edges no longer fade
 - Power drain now ramps at 0.1/s beginning at 30 seconds